remove dynamic client usage; go mod tidy/vendor
