+++ pkg/shp/cmd/buildrun/cancel.go
@@ -1,17 +1,19 @@
 package buildrun
 
 import (
+	"encoding/json"
 	"fmt"
 
 	"github.com/spf13/cobra"
 
 	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 
 	buildv1alpha1 "github.com/shipwright-io/build/pkg/apis/build/v1alpha1"
 	"github.com/shipwright-io/cli/pkg/shp/cmd/runner"
 	"github.com/shipwright-io/cli/pkg/shp/params"
-	"github.com/shipwright-io/cli/pkg/shp/resource"
 )
 
 // CancelCommand contains data input from user for delete sub-command
@@ -50,10 +52,13 @@ func (c *CancelCommand) Validate() error {
 
 // Run executes cancel sub-command logic
 func (c *CancelCommand) Run(params *params.Params, ioStreams *genericclioptions.IOStreams) error {
-	brr := resource.GetBuildRunResource(params)
+	clientset, err := params.ShipwrightClientSet()
+	if err != nil {
+		return err
+	}
 
 	br := &buildv1alpha1.BuildRun{}
-	if err := brr.Get(c.cmd.Context(), c.name, br); err != nil {
+	if br, err = clientset.ShipwrightV1alpha1().BuildRuns(params.Namespace()).Get(c.cmd.Context(), c.name, metav1.GetOptions{}); err != nil {
 		return fmt.Errorf("failed to retrieve BuildRun %s: %s", c.name, err.Error())
 	}
 	//TODO replace with br.IsDone() when that is available and vendored in
@@ -63,7 +68,21 @@ func (c *CancelCommand) Run(params *params.Params, ioStreams *genericclioptions.
 	}
 
 	//TODO use constant when vendor in api changes
-	if err := brr.Patch(c.cmd.Context(), c.name, "replace", "/spec/state", "BuildRunCanceled"); err != nil {
+	type patchStringValue struct {
+		Op    string `json:"op"`
+		Path  string `json:"path"`
+		Value string `json:"value"`
+	}
+	payload := []patchStringValue{{
+		Op:    "replace",
+		Path:  "/spec/state",
+		Value: buildv1alpha1.BuildRunStateCancel,
+	}}
+	var data []byte
+	if data, err = json.Marshal(payload); err != nil {
+		return err
+	}
+	if _, err = clientset.ShipwrightV1alpha1().BuildRuns(params.Namespace()).Patch(c.Cmd().Context(), c.name, types.JSONPatchType, data, metav1.PatchOptions{}); err != nil {
 		return err
 	}
