--- pkg/shp/resource/doc.go
@@ -1,4 +1,0 @@
-// Package resource contains Resource type and functions for interacting with it's a wrapper around
-// kubernetes dynamic client that has most of the necessary type info encapsulated which allows us to
-// manipulate Kubernetes objects.
-package resource
